# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Student Leader Cabinet API
  version: 0.6.0
  description: >
    REST API личного кабинета студенческого лидера. Все ответы обёрнуты в envelope
    {data, meta, request_id} либо {error, request_id} (SITE.md §20, §50).
servers:
  - url: https://eazytech.ru/api/v1
    description: production
  - url: http://localhost:8080/api/v1
    description: local
paths:
...
  /events/{eventSlug}/participant-auth/fio:
    parameters: [{ $ref: '#/components/parameters/EventSlug' }]
    post:
      tags: [Participant auth]
      summary: Вход участника по ФИО и дате рождения
      operationId: participantLoginByName
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParticipantNameLogin' }
      responses:
        '200': { $ref: '#/components/responses/ParticipantSession' }
        default: { $ref: '#/components/responses/Error' }
  /events/{eventSlug}/participant-auth/union-card:
    parameters: [{ $ref: '#/components/parameters/EventSlug' }]
    post:
      tags: [Participant auth]
      summary: Вход участника по номеру профсоюзного билета
      operationId: participantLoginByUnionCard
      requestBody:
        $ref: '#/components/requestBodies/ParticipantIdentifier'
      responses:
        '200': { $ref: '#/components/responses/ParticipantSession' }
        default: { $ref: '#/components/responses/Error' }
  /events/{eventSlug}/participant-auth/sks:
    parameters: [{ $ref: '#/components/parameters/EventSlug' }]
    post:
      tags: [Participant auth]
      summary: Вход участника по barcode СКС
      operationId: participantLoginBySKS
      requestBody:
        $ref: '#/components/requestBodies/ParticipantIdentifier'
      responses:
        '200': { $ref: '#/components/responses/ParticipantSession' }
        default: { $ref: '#/components/responses/Error' }
  /participant/me:
    get:
      tags: [Participant]
      summary: Текущий участник и мероприятие
      operationId: participantMe
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/ParticipantSession' }
        default: { $ref: '#/components/responses/Error' }
  /participant/logout:
    post:
      tags: [Participant auth]
      summary: Завершить текущую participant session
      operationId: participantLogout
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/points:
    get:
      tags: [Points]
      summary: Баланс и журнал баллов текущего участника
      operationId: participantPoints
      security: [{ participantCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/lectures:
    get:
      tags: [Attendance]
      summary: Лекции и посещения текущего участника
      operationId: participantLectures
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/qr:
    post:
      tags: [Attendance]
      summary: Выпустить одноразовый QR-код посещения
      operationId: participantIssueQR
      security: [{ participantCookie: [] }]
      responses:
        '201': { $ref: '#/components/responses/Created' }
        default: { $ref: '#/components/responses/Error' }
  /participant/tasks:
    get:
      tags: [Tasks]
      summary: Активные задания и отправки текущего участника
      operationId: participantTasks
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/tasks/{taskId}:
    parameters: [{ $ref: '#/components/parameters/TaskId' }]
    get:
      tags: [Tasks]
      summary: Задание текущего участника
      operationId: participantTask
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/tasks/{taskId}/submissions:
    parameters: [{ $ref: '#/components/parameters/TaskId' }]
    post:
      tags: [Tasks]
      summary: Отправить подтверждение выполнения задания
      operationId: participantSubmitTask
      security: [{ participantCookie: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                participant_comment: { type: string, maxLength: 2000 }
                links: { type: array, maxItems: 10, items: { type: string, format: uri } }
                images: { type: array, maxItems: 10, items: { type: string, format: binary } }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        default: { $ref: '#/components/responses/Error' }
  /participant/task-assets/{assetId}:
    parameters: [{ $ref: '#/components/parameters/AssetId' }]
    get:
      tags: [Tasks]
      summary: Получить временную ссылку на собственный файл задания
      operationId: participantTaskAsset
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/DownloadURL' }
        default: { $ref: '#/components/responses/Error' }
  /participant/merch:
    get:
      tags: [Merch]
      summary: Доступный каталог мерча
      operationId: participantMerch
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/merch/{productSlug}:
    parameters: [{ $ref: '#/components/parameters/ProductSlug' }]
    get:
      tags: [Merch]
      summary: Карточка товара
      operationId: participantMerchProduct
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/merch-saving-target:
    put:
      tags: [Merch]
      summary: Установить цель накопления
      operationId: participantSetSavingTarget
      security: [{ participantCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id]
              properties: { product_id: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Merch]
      summary: Удалить цель накопления
      operationId: participantDeleteSavingTarget
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/orders:
    get:
      tags: [Merch]
      summary: Заказы текущего участника
      operationId: participantOrders
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Merch]
      summary: Зарезервировать товары и баллы
      operationId: participantReserveOrder
      security: [{ participantCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveOrderInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '201': { $ref: '#/components/responses/Created' }
        default: { $ref: '#/components/responses/Error' }
  /participant/orders/{orderId}:
    parameters: [{ $ref: '#/components/parameters/OrderId' }]
    get:
      tags: [Merch]
      summary: Собственный заказ
      operationId: participantOrder
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
  /participant/orders/{orderId}/cancel:
    parameters: [{ $ref: '#/components/parameters/OrderId' }]
    post:
      tags: [Merch]
      summary: Отменить зарезервированный заказ
      operationId: participantCancelOrder
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '201': { $ref: '#/components/responses/Created' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  parameters:
...
    TaskId: { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    AssetId: { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
...
    ProductSlug: { name: productSlug, in: path, required: true, schema: { type: string } }
...
    OrderId: { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
    EventSlug: { name: eventSlug, in: path, required: true, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    Offset: { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 200 }
  requestBodies:
    ParticipantIdentifier:
      required: true
      content:
        application/json:
          schema:
            type: object
            description: Допустимы value либо профильное имя поля endpoint.
            properties:
              value: { type: string }
              union_card_number: { type: string }
              sks_barcode: { type: string }
  responses:
    Success:
      description: Успешный ответ в общем envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Created:
      description: Ресурс создан; идемпотентный replay может вернуть 200
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    ParticipantSession:
      description: Participant cookie установлена или уже подтверждена
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/SuccessEnvelope' }
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/ParticipantSession' }
    DownloadURL:
      description: Короткоживущая ссылка на объект
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/SuccessEnvelope' }
              - type: object
                properties:
                  data:
                    type: object
                    required: [download_url]
                    properties: { download_url: { type: string, format: uri } }
    Error:
      description: Ошибка в общем envelope; HTTP status и error.code зависят от причины
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties:
        data: {}
        meta:
          type: object
          nullable: true
        request_id:
          type: string
          format: uuid
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTH_INVALID_CREDENTIALS
                - AUTH_SESSION_EXPIRED
                - AUTH_REFRESH_REUSED
                - AUTH_ACCOUNT_BLOCKED
                - AUTH_PASSWORD_CHANGE_REQUIRED
                - FORBIDDEN
                - CONTEST_ACCESS_DENIED
                - RESOURCE_NOT_FOUND
                - SUBMISSION_LOCKED
                - SUBMISSION_DEADLINE_PASSED
                - SUBMISSION_VERSION_CONFLICT
                - SUBMISSION_ALREADY_SUBMITTED
                - UPLOAD_NOT_COMPLETED
                - FILE_TYPE_NOT_ALLOWED
                - FILE_TOO_LARGE
                - IDEMPOTENCY_CONFLICT
                - RATE_LIMIT_EXCEEDED
                - TELEGRAM_DELIVERY_FAILED
                - NOT_FOUND
                - PARTICIPANT_IDENTIFIER_TAKEN
                - PARTICIPANT_IDENTITY_AMBIGUOUS
                - PARTICIPANT_AUTH_FAILED
                - PARTICIPANT_SESSION_EXPIRED
                - LECTURE_NOT_FOUND
                - INVALID_LECTURE_TRANSITION
                - LECTURE_HAS_ATTENDANCE
                - ATTENDANCE_CLOSED
                - QR_INVALID
                - QR_EXPIRED
                - QR_REPLAYED
                - PARTICIPANT_UNAVAILABLE
                - TASK_NOT_FOUND
                - INVALID_TASK_TRANSITION
                - TASK_SUBMISSION_CLOSED
                - STORAGE_UNAVAILABLE
                - MERCH_NOT_FOUND
                - INVALID_MERCH_TRANSITION
                - INSUFFICIENT_STOCK
                - INSUFFICIENT_POINTS
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
              nullable: true
        request_id:
          type: string
...
    ParticipantStatus: { type: string, enum: [ACTIVE, BLOCKED, ARCHIVED] }
...
    Participant:
      type: object
      required: [id, event_id, full_name, birth_date, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        full_name: { type: string }
        birth_date: { type: string, format: date }
        union_card_number: { type: string, nullable: true }
        sks_barcode: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ParticipantStatus' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: string, format: date-time, nullable: true }
    EventRef:
      type: object
      required: [id, slug, name, timezone]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        timezone: { type: string }
    ParticipantSession:
      type: object
      required: [participant, event]
      properties:
        participant: { $ref: '#/components/schemas/Participant' }
        event: { $ref: '#/components/schemas/EventRef' }
        expires_at: { type: string, format: date-time }
    ParticipantNameLogin:
      type: object
      required: [full_name, birth_date]
      properties:
        full_name: { type: string }
        birth_date: { type: string, format: date }
...
    ReserveOrderInput:
      type: object
      required: [items]
      properties:
        idempotency_key: { type: string, minLength: 8, maxLength: 200 }
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    participantCookie:
      type: apiKey
      in: cookie
      name: slc_participant_session